+++ packages/headless/src/controllers/facets/automatic-facet/headless-automatic-facet.ts
@@ -3,8 +3,14 @@ import {
   deselectAllAutomaticFacetValues,
   toggleSelectAutomaticFacetValue,
 } from '../../../features/facets/automatic-facet-set/automatic-facet-set-actions';
-import {logFacetClearAll} from '../../../features/facets/facet-set/facet-set-analytics-actions';
-import {getAnalyticsActionForToggleFacetSelect} from '../../../features/facets/facet-set/facet-set-utils';
+import {
+  facetClearAll,
+  logFacetClearAll,
+} from '../../../features/facets/facet-set/facet-set-analytics-actions';
+import {
+  getLegacyAnalyticsActionForToggleFacetSelect,
+  getAnalyticsActionForToggleFacetSelect,
+} from '../../../features/facets/facet-set/facet-set-utils';
 import {FacetValue} from '../../../features/facets/facet-set/interfaces/response';
 import {executeSearch} from '../../../features/search/search-actions';
 import {buildController} from '../../controller/headless-controller';
@@ -47,14 +53,23 @@ export function buildAutomaticFacet(
       dispatch(toggleSelectAutomaticFacetValue({field, selection}));
       dispatch(
         executeSearch({
-          legacy: getAnalyticsActionForToggleFacetSelect(field, selection),
+          legacy: getLegacyAnalyticsActionForToggleFacetSelect(
+            field,
+            selection
+          ),
+          next: getAnalyticsActionForToggleFacetSelect(field, selection),
         })
       );
     },
 
     deselectAll() {
       dispatch(deselectAllAutomaticFacetValues(field));
-      dispatch(executeSearch({legacy: logFacetClearAll(field)}));
+      dispatch(
+        executeSearch({
+          legacy: logFacetClearAll(field),
+          next: facetClearAll(field),
+        })
+      );
     },
 
     get state() {