+++ packages/headless/src/controllers/facets/category-facet/headless-category-facet-search.ts
@@ -3,7 +3,10 @@ import {SearchThunkExtraArguments} from '../../../app/search-thunk-extra-argumen
 import {updateFacetOptions} from '../../../features/facet-options/facet-options-actions';
 import {registerCategoryFacetSearch} from '../../../features/facets/facet-search-set/category/category-facet-search-actions';
 import {defaultFacetSearchOptions} from '../../../features/facets/facet-search-set/facet-search-reducer-helpers';
-import {logFacetSelect} from '../../../features/facets/facet-set/facet-set-analytics-actions';
+import {
+  facetSelect,
+  logFacetSelect,
+} from '../../../features/facets/facet-set/facet-set-analytics-actions';
 import {executeSearch} from '../../../features/search/search-actions';
 import {
   CategoryFacetSearchSection,
@@ -53,6 +56,7 @@ export function buildCategoryFacetSearch(
       dispatch(
         executeSearch({
           legacy: logFacetSelect({facetId, facetValue: value.rawValue}),
+          next: facetSelect(facetId, value.rawValue),
         })
       );
     },