+++ packages/headless/src/controllers/field-suggestions/facet/headless-field-suggestions.ts
@@ -6,6 +6,8 @@ import {updateFacetOptions} from '../../../features/facet-options/facet-options-
 import {specificFacetSearchSetReducer as facetSearchSet} from '../../../features/facets/facet-search-set/specific/specific-facet-search-set-slice';
 import {registerFacet} from '../../../features/facets/facet-set/facet-set-actions';
 import {
+  facetExclude,
+  facetSelect,
   logFacetExclude,
   logFacetSelect,
 } from '../../../features/facets/facet-set/facet-set-analytics-actions';
@@ -182,6 +184,7 @@ export function buildFieldSuggestions(
       engine.dispatch(
         executeSearch({
           legacy: logFacetSelect({facetId, facetValue: value.rawValue}),
+          next: facetSelect(facetId, value.rawValue),
         })
       );
     },
@@ -190,6 +193,7 @@ export function buildFieldSuggestions(
       engine.dispatch(
         executeSearch({
           legacy: logFacetExclude({facetId, facetValue: value.rawValue}),
+          next: facetExclude(facetId, value.rawValue),
         })
       );
     },