+++ packages/headless/src/features/facets/facet-set/facet-set-slice.test.ts
@@ -784,7 +784,7 @@ describe('facet-set slice', () => {
 
     it.each([
       {hash: 'f', facetValueState: 'selected' as FacetValueState},
-      {hash: 'f-excluded', facetValueState: 'excluded' as FacetValueState},
+      {hash: 'fExcluded', facetValueState: 'excluded' as FacetValueState},
     ])(
       `when a facet is not found in the #$hash payload,
     it deselects all values by setting the state of each facet value in #currentValues to idle`,