+++ packages/headless/src/features/facets/facet-set/facet-set-utils.ts
@@ -1,4 +1,8 @@
+import {SearchAction} from '../../search/search-actions';
 import {
+  facetDeselect,
+  facetExclude,
+  facetSelect,
   logFacetDeselect,
   logFacetExclude,
   logFacetSelect,
@@ -14,7 +18,7 @@ export const isFacetValueExcluded = (value: FacetValue) => {
   return value.state === 'excluded';
 };
 
-export const getAnalyticsActionForToggleFacetSelect = (
+export const getLegacyAnalyticsActionForToggleFacetSelect = (
   facetId: string,
   selection: FacetValue
 ) => {
@@ -28,7 +32,16 @@ export const getAnalyticsActionForToggleFacetSelect = (
     : logFacetSelect(payload);
 };
 
-export const getAnalyticsActionForToggleFacetExclude = (
+export const getAnalyticsActionForToggleFacetSelect = (
+  facetId: string,
+  selection: FacetValue
+): SearchAction => {
+  return isFacetValueSelected(selection)
+    ? facetDeselect(facetId, selection.value)
+    : facetSelect(facetId, selection.value);
+};
+
+export const getLegacyAnalyticsActionForToggleFacetExclude = (
   facetId: string,
   selection: FacetValue
 ) => {
@@ -40,4 +53,13 @@ export const getAnalyticsActionForToggleFacetExclude = (
   return isFacetValueExcluded(selection)
     ? logFacetDeselect(payload)
     : logFacetExclude(payload);
+};
+
+export const getAnalyticsActionForToggleFacetExclude = (
+  facetId: string,
+  selection: FacetValue
+): SearchAction => {
+  return isFacetValueExcluded(selection)
+    ? facetDeselect(facetId, selection.value)
+    : facetExclude(facetId, selection.value);
 };